[browsertime] WIP: Add support for "privileged" measurements.
This runs measurement scripts in Firefox's privileged "chrome"
JavaScript context (really, Gecko's context).  That is roughly
equivalent to executing in the browser console.

This patch is WIP for a few reasons:

1) the filtering is done based on file name for convenience.  We can
filter in many better ways, at least by script category to start.

2) this will break in non Firefox/Gecko browsers.  We can just drop
these measurements in other browsers, but that isn't done yet.